+++ lib/src/extensions/date_extension.dart
@@ -1,0 +1,7 @@
+/// Extension on [DateTime] to add more functionality.
+extension FormxDateExtension on DateTime {
+  /// Whether this date has passed the given [duration] in time.
+  bool hasPassed(Duration duration) {
+    return isBefore(DateTime.now().subtract(duration));
+  }
+}